+++ include/Timing.hpp
@@ -12,15 +12,15 @@ class Timing
 
     struct TimingResult
     {
-        std::chrono::duration<double> avg;
-        std::chrono::duration<double> eta;
-        std::chrono::duration<double, std::milli> epoch_ms;
+        std::chrono::duration<float> avg;
+        std::chrono::duration<float> eta;
+        std::chrono::duration<float, std::milli> epoch_ms;
     };
 
     TimingResult update(std::chrono::steady_clock::duration epoch_time, int epochs_left);
 
   private:
-    std::chrono::duration<double> sum_recent_times{};
+    std::chrono::duration<float> sum_recent_times{};
     std::deque<std::chrono::steady_clock::duration> recent_times{};
 
     size_t window_size;
