+++ src/Timing.cpp
@@ -15,11 +15,11 @@ Timing::TimingResult Timing::update(std::chrono::steady_clock::duration epoch_ti
         recent_times.pop_back();
     }
 
-    std::chrono::duration<double> avg_epoch_time = sum_recent_times / recent_times.size();
-    std::chrono::duration<double> eta = epochs_left * avg_epoch_time;
+    std::chrono::duration<float> avg_epoch_time = sum_recent_times / recent_times.size();
+    std::chrono::duration<float> eta = epochs_left * avg_epoch_time;
 
-    std::chrono::duration<double, std::milli> epoch_ms =
-        std::chrono::duration_cast<std::chrono::duration<double, std::milli>>(epoch_time);
+    std::chrono::duration<float, std::milli> epoch_ms =
+        std::chrono::duration_cast<std::chrono::duration<float, std::milli>>(epoch_time);
 
     return Timing::TimingResult{avg_epoch_time, eta, epoch_ms};
 }
